# std.uni.isPunctuation: significant performance increase for ASCII