With the test character `dchar a = '.';` with one million calls:

```
isPunctuation		79 ms, 385 μs, and 4 hnsecs
isPunctuation2		7 ms, 58 μs, and 9 hnsecs
```