@eco 2016-04-04
What's the performance cost for the non-ASCII case?

It seems like we could avoid the branch by using the type system (i.e. specialize for `std.encoding.AsciiString`) though if this just makes the vast majority of punctuation faster (I know nothing about non-English punctuation) it might as well be considered unicode optimization.