@JackStouffer 2016-04-04
> What's the performance cost for the non-ASCII case?

11x slower

> It seems like we could avoid the branch by using the type system

How would that work?

> though if this just makes the vast majority of punctuation faster (I know nothing about non-English punctuation) it might as well be considered unicode optimization

Not the vast majority, unicode supports something like 500+ different punctuation symbols.