@eco 2016-04-04
I mean, how much does the added branch hurt performance for the non-ASCII case? I imagine it's very tiny; just curious if you had tested.

> How would that work?

Sorry, just kind of wondering aloud when I saw the added branch about how it could be avoided. I forgot this was dealing with dchar, not string. The pull request looks good though @DmitryOlshansky should probably weigh in.