@JackStouffer 2016-04-04
> I mean, how much does the added branch hurt performance for the non-ASCII case? I imagine it's very tiny; just curious if you had tested.

I hadn't. `isASCII` is one `if`statement, so if DMD is inlining correctly the performance should be `<= 1ns`.