@DmitryOlshansky 2016-04-05
@JackStouffer One last thing to check is please benchmark your latest version with isASCII check as function.