@JackStouffer 2016-04-05
@DmitryOlshansky fixed

```
isPunctuation		77 ms, 169 μs, and 3 hnsecs
isPunctuation2		9 ms, 698 μs, and 8 hnsecs
```