@DmitryOlshansky 2016-04-06
So somehow it is slower, I suggest to just use < 0x80 in such a case.