@JackStouffer 2016-04-06
> So somehow it is slower, I suggest to just use < 0x80 in such a case.

That's only because I am testing the numbers with the debug make build, and therefore nothing is inlined, because it's the only thing that I can get my dmd set up to compile.