@DmitryOlshansky 2016-04-06
Thing is std.ascii always had problem with inlining... Anyhow you'd better fix your dmd setup