@JackStouffer 2016-04-06
@DmitryOlshansky Because I have no way to prove that the call will be inlined, I changed it to the manual check.