@DmitryOlshansky 2016-04-11
Auto-merge toggled on